Don't stop flushing until successful
Previously, when the input stream of requests terminated, clients would
only try to call `poll_shutdown` _once_. Regardless of whether that call
succeeded or failed, it would set `rx_only`, which disables further
flushing. Perhaps unsurprisingly, this is a problem, as there may still
be data that needs to be written out.