Do not panic on server desynchronization
Previously, we would panic if the server sends a response to a request
the client does not have a record of. With this change, it is up to the
user whether this should cause a panic or not. We raise it as a new
error, `Error::Desynchronized`, and the user can to panic in the error
handler if they so wish.

This is a backwards-incompatible change.

Fixes #15.